Repair dockerfile, docker-compose add loadusers and loadbooks method
