+++ docker-compose.yml
@@ -10,12 +10,13 @@ services:
       MYSQL_DATABASE: library
   JJDZR11-Konstruktorsi:
     image: library:0.1
+    restart: always
     depends_on:
       - mysql
     ports:
       - 8080:8080
     environment:
-      SPRING_DATASOURCE_URL: jdbc:mysql://mysql:3308/library
+      SPRING_DATASOURCE_URL: jdbc:mysql://mysql:3306/library
       SPRING_DATASOURCE_USERNAME: root
       SPRING_DATASOURCE_PASSWORD: book
       SPRING_DATASOURCE_DRIVER-CLASS-NAME: com.mysql.cj.jdbc.Driver
