+++ src/main/java/pl/isa/biblioteka/config/SpringSecurityConfig.java
@@ -23,7 +23,7 @@ class SpringSecurityConfig {
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http.authorizeRequests(request -> request
-                        .requestMatchers("/", "/images/**", "/css/**", "/static/font/**", "/font/**", "/searchText", "/searchByText", "/list", "/bookList/**", "/register", "/top/**").permitAll()
+                        .requestMatchers("/", "/images/**", "/css/**", "/static/font/**", "/font/**", "/searchText", "/searchByText", "/list", "/bookList/**", "/register", "/top/**", "/loadusers").permitAll()
                         .anyRequest().authenticated())
                 .formLogin(form -> form
                         .loginPage("/").usernameParameter("user").passwordParameter("password")
